% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_bracket}
\alias{add_bracket}
\title{Add a single bracket grob to a ggplot}
\usage{
add_bracket(p, grob)
}
\arguments{
\item{p}{A ggplot object.}

\item{grob}{A grob from \code{*_grob()} functions.}
}
\value{
A ggplot with the annotation added.
}
\description{
Convenience wrapper around \code{ggplot2::annotation_custom()}.
}
