% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_by_groups}
\alias{add_by_groups}
\title{Add bracket(s) by group indices (any side)}
\usage{
add_by_groups(
  p,
  side = c("top", "bottom", "left", "right"),
  labels,
  n_groups,
  g_start,
  g_end = g_start,
  pos = if (match.arg(side, c("top", "bottom", "left", "right")) \%in\% c("top",
    "right")) 1.06 else -0.06,
  pad = 0.02,
  margin_pt = 60,
  ...
)
}
\description{
Add bracket(s) by group indices (any side)
}
