% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_faceted_brackets}
\alias{add_faceted_brackets}
\title{Add facet-targeted brackets using a specs df (rows matched to facet columns)}
\usage{
add_faceted_brackets(
  p,
  specs_df,
  start_col = "start",
  end_col = "end",
  pos_col = "pos",
  label_col = "label",
  side_col = "side",
  margin_pt = 60,
  ...
)
}
\description{
Add facet-targeted brackets using a specs df (rows matched to facet columns)
}
