% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_horizontal_bot}
\alias{add_horizontal_bot}
\title{Add a bottom horizontal bracket to a ggplot (with auto margin)}
\usage{
add_horizontal_bot(p, label, xstart, xend, y, margin_pt = 40, ...)
}
\arguments{
\item{p}{A ggplot object.}

\item{label}{Text label.}

\item{xstart, xend}{Horizontal start/end (0–1); @param y Vertical NPC (0–1).}

\item{margin_pt}{Minimum bottom margin in points. Default \code{40}.}
}
\description{
Add a bottom horizontal bracket to a ggplot (with auto margin)
}
