% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_horizontal_top_grob}
\alias{add_horizontal_top_grob}
\title{Top horizontal bracket grob}
\usage{
add_horizontal_top_grob(label, xstart, xend, y, ...)
}
\arguments{
\item{label}{Character label to draw.}

\item{xstart, xend}{Horizontal start/end (0–1); @param y Vertical NPC (0–1).}
}
\value{
A grid grob.
}
\description{
Top horizontal bracket grob
}
