% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_vertical_left_grob}
\alias{add_vertical_left_grob}
\title{Vertical bracket grob on the left side}
\usage{
add_vertical_left_grob(label, ystart, yend, x, ...)
}
\arguments{
\item{label}{Text label.}

\item{ystart, yend}{Vertical start/end (0–1); @param x Horizontal NPC (0–1).}
}
\value{
A grid grob.
}
\description{
Vertical bracket grob on the left side
}
