% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_vertical_right}
\alias{add_vertical_right}
\title{Add a right vertical bracket to a ggplot (with auto margin)}
\usage{
add_vertical_right(p, label, ystart, yend, x, margin_pt = 40, ...)
}
\arguments{
\item{p}{A ggplot object.}

\item{label}{Text label.}

\item{ystart, yend, x}{See \code{\link{add_vertical_right_grob}}.}

\item{margin_pt}{Minimum right margin in points. Default \code{40}.}
}
\description{
Add a right vertical bracket to a ggplot (with auto margin)
}
