% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{npc_from_groups}
\alias{npc_from_groups}
\title{Compute NPC start/end from group indices (equal-width bins)}
\usage{
npc_from_groups(
  n_groups,
  g_start,
  g_end = g_start,
  along = c("x", "y"),
  pad = 0
)
}
\description{
Compute NPC start/end from group indices (equal-width bins)
}
