% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{set_pmplot_theme}
\alias{set_pmplot_theme}
\title{Apply pmplot theme globally}
\usage{
set_pmplot_theme(base_family = NULL, base_size = 10, base_face = "plain")
}
\description{
Apply pmplot theme globally
}
