% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_brackets}
\alias{add_brackets}
\title{Add multiple horizontal and/or vertical brackets in one call (specs df)
Expected columns: side, start, end, pos, label}
\usage{
add_brackets(p, specs_df, base_pt = 40, step_pt = 14, ...)
}
\description{
Add multiple horizontal and/or vertical brackets in one call (specs df)
Expected columns: side, start, end, pos, label
}
