% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{theme_pmplot}
\alias{theme_pmplot}
\title{Minimal print-friendly theme (base_family NULL by default)}
\usage{
theme_pmplot(base_family = NULL, base_size = 10, base_face = "plain")
}
\description{
Minimal print-friendly theme (base_family NULL by default)
}
